+++ packages/core/src/extraTypes.ts
@@ -1,0 +1,3 @@
+export type UnknownArgs = unknown[] | [];
+
+export type Func<Args extends UnknownArgs, ReturnType = void> = (...args: Args) => ReturnType;