+++ packages/core/tests/effect.test.ts
@@ -1,4 +1,4 @@
-import { store, effect, StoreMiddleware } from "..";
+import { store, effect, StoreMiddleware } from "../src";
 import * as chai from "chai";
 
 const { expect } = chai;
@@ -6,7 +6,7 @@ const { expect } = chai;
 describe("effect()", () => {
   const initialState = { foo: "bar", baz: [1, 2, 3], bar: "bar" };
   const storeName = "store-name";
-  const middlewares: StoreMiddleware<typeof initialState>[] = [storeManager => storeManager];
+  const middlewares: StoreMiddleware<typeof initialState>[] = [store => store];
 
   let testStore: typeof initialState;
 
@@ -43,8 +43,33 @@ describe("effect()", () => {
     expect(testStore).to.deep.equal({ ...initialState, bar: "22" });
   });
 
+  it("should be able to update store via actions combination (deep)", () => {
+    const partialUpdate = effect(testStore);
+    const updateBar = effect(testStore, "bar");
+    const updateFoo = effect(testStore, "foo");
+    const updateBaz = effect(testStore, (baz?: number[]) => partialUpdate({ baz, foo: testStore.foo + testStore.bar }));
+
+    updateBar("barbarian");
+    updateFoo("foobarbarian");
+    updateBaz([6, 6, 6]);
+    partialUpdate({ baz: [...testStore.baz, testStore.bar.length] });
+    expect(testStore).to.deep.equal({
+      ...initialState,
+      bar: "barbarian",
+      foo: "foobarbarianbarbarian",
+      baz: [6, 6, 6, "barbarian".length],
+    });
+  });
+
   it("should be able to work with async actions", async () => {
-    await effect(testStore, async () => ({ bar: "23" }))();
+    const updateBar = effect(testStore, async () => ({ bar: "23" }));
+    await updateBar();
+    expect(testStore).to.deep.equal({ ...initialState, bar: "23" });
+  });
+
+  it("should be able to work with async actions (+input params)", async () => {
+    const updateBar = effect(testStore, async (bar: string) => ({ bar }));
+    await updateBar("23");
     expect(testStore).to.deep.equal({ ...initialState, bar: "23" });
   });
 });