+++ packages/core/tests/store.test.ts
@@ -1,5 +1,9 @@
-import { store, StoreMiddleware } from "..";
-import { expect } from "chai";
+import { store, StoreMiddleware, StoreType, manage } from "../src";
+import * as chai from "chai";
+import * as spies from "chai-spies";
+chai.use(spies);
+
+const { spy, expect } = chai;
 
 describe("store()", () => {
   const initialState = { foo: "bar", baz: [1, 2, 3], bar: "bar" };
@@ -25,6 +29,36 @@ describe("store()", () => {
       it("return value should be equal to initial state", () => {
         expect(testStore).to.deep.equal(initialState);
       });
+    });
+  });
+
+  describe("middlewares", () => {
+    it("should call middleware when create a store", () => {
+      const middlewareSpy = spy();
+      const middleware = <Store extends StoreType>(store: Store) => {
+        middlewareSpy();
+        return store;
+      };
+      store({ foo: "bar", bar: "baz" }, storeName, [middleware]);
+      expect(middlewareSpy).to.been.called.once;
+    });
+    it("should be able to mutate store manager", () => {
+      const middleware = <Store extends StoreType>(store: Store) => {
+        const manager = manage(store);
+        manager.partialUpdate = () => ({});
+        return store;
+      };
+      const testStore = store({ foo: "bar", bar: "baz" }, storeName, [middleware]);
+      manage(testStore).partialUpdate({ foo: "baz" });
+      expect(testStore).to.deep.equal({ foo: "bar", bar: "baz" });
+    });
+    it("should be able to store state", () => {
+      const middleware = <Store extends StoreType>(store: Store) => ({
+        ...store,
+        newProp: "value",
+      });
+      const testStore = store({ foo: "bar", bar: "baz" }, storeName, [middleware]);
+      expect(testStore).to.deep.equal({ foo: "bar", bar: "baz", newProp: "value" });
     });
   });
 });